requirements.txt
aniso8601==8.1.0
astroid==2.4.2
CacheControl==0.12.6
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
firebase-admin==4.5.1
firestore==0.0.8
Flask==1.1.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-bigquery==2.6.2
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
gunicorn==19.9.0
httplib2==0.18.1
idna==2.10
iso8601==0.1.13
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
rsa==4.7
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==0.16.0
wrapt==1.12.1
